requirements/dev.txt
-r common.txt
django-rest-framework-docs==0.1.7
